split serializer/deserializer
