+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/internal/serialization/deserializers/SignalDeserializer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.disk.buffering.internal.serialization.serializers;
+package io.opentelemetry.contrib.disk.buffering.internal.serialization.deserializers;
 
 import io.opentelemetry.sdk.logs.data.LogRecordData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
@@ -13,15 +13,15 @@ import java.util.List;
 public interface SignalDeserializer<SDK_ITEM> {
 
   static SignalDeserializer<SpanData> ofSpans() {
-    return SpanDataSerializer.getInstance();
+    return SpanDataDeserializer.getInstance();
   }
 
   static SignalDeserializer<MetricData> ofMetrics() {
-    return MetricDataSerializer.getInstance();
+    return MetricDataDeerializer.getInstance();
   }
 
   static SignalDeserializer<LogRecordData> ofLogs() {
-    return LogRecordDataSerializer.getInstance();
+    return LogRecordDataDeserializer.getInstance();
   }
 
   List<SDK_ITEM> deserialize(byte[] source);
